chore: rename downloader module to scraper
